Switch to Numeric

+++ src/controller/src/clusters.rs
@@ -40,6 +40,7 @@ use mz_repr::GlobalId;
 use crate::Controller;
 
 pub use mz_compute_client::controller::DEFAULT_COMPUTE_REPLICA_LOGGING_INTERVAL_MICROS as DEFAULT_REPLICA_LOGGING_INTERVAL_MICROS;
+use mz_repr::adt::numeric::Numeric;
 
 /// Identifies a cluster.
 pub type ClusterId = ComputeInstanceId;
@@ -80,7 +81,7 @@ pub struct ReplicaAllocation {
     /// The number of worker threads in the replica.
     pub workers: usize,
     /// The number of compute credits per hour.
-    pub compute_credits_per_hour: f64,
+    pub compute_credits_per_hour: Numeric,
 }
 
 /// Configures the location of a cluster replica.
